feat: reorganize search component structure and remove unused types and files
