+++ components/search/SearchInput.vue
@@ -1,16 +1,14 @@
 <script setup lang="ts">
-import { ref, computed, watch, nextTick } from 'vue'
+import { watch, computed } from 'vue'
+import { useQualifiers } from './composables/useQualifiers'
+import { useSearchParser } from './composables/useSearchParser'
+import { useScrollLock } from './composables/useScrollLock'
 import { useSearchInput } from './composables/useSearchInput'
-import { useSearchSuggestions } from './composables/useSearchSuggestions'
-import { useSearchShortcuts } from './composables/useSearchShortcuts'
-import SearchDialog from './SearchDialog.vue'
-import SearchHighlight from './SearchHighlight.vue'
-import SearchSuggestions from './SearchSuggestions.vue'
-import type { SearchMode } from './types'
+import { useSearchInteractions } from './composables/useSearchInteractions'
 
 const props = defineProps<{
   modelValue: string
-  mode?: SearchMode
+  mode?: 'inline' | 'modal'
 }>()
 
 const emit = defineEmits<{
@@ -18,313 +16,361 @@ const emit = defineEmits<{
   (e: 'search'): void
 }>()
 
-const {
-  searchInput,
-  input,
-  cursorPosition,
-  isInputFocused,
-  coloredParts,
-  parseSearchTokens,
-  updateCursorPosition
-} = useSearchInput()
+// 1. Primero inicializamos los composables base
+const { QUALIFIERS } = useQualifiers()
+const { parseSearchTokens, colorText } = useSearchParser()
+const { lockScroll, unlockScroll } = useScrollLock()
 
-const tokens = computed(() => parseSearchTokens(searchInput.value))
-const {
-  showSuggestions,
-  suggestions,
-  handleSuggestionSelect,
-  handleSuggestionRemove
-} = useSearchSuggestions(
-  () => tokens.value,
-  () => searchInput.value
+// 1. Crear una función handleFocus que usaremos más tarde
+const handleFocusWrapper = () => {
+  searchInteractions.handleFocus()
+}
+
+// 2. Inicializar useSearchInput primero, pasando la función wrapper
+const searchInput = useSearchInput(handleFocusWrapper)
+
+// 3. Inicializar useSearchInteractions después
+const searchInteractions = useSearchInteractions(
+  props,
+  emit,
+  searchInput.input,
+  searchInput.updateCursorPosition,
+  lockScroll,
+  unlockScroll
 )
 
-// Estado local
-const isExpanded = ref(false)
-const originalPosition = ref(0)
-const isInlineMode = computed(() => props.mode === 'inline')
+// 4. Desestructurar los valores que necesitamos de ambos composables
+const {
+  input,
+  cursorPosition,
+  searchContainer,
+  isInputFocused,
+  handleInputScroll
+} = searchInput
 
-// Funciones de manejo de scroll
-const lockScroll = () => {
-  originalPosition.value = window.scrollY
-  document.body.style.position = 'fixed'
-  document.body.style.top = `-${originalPosition.value}px`
-  document.body.style.width = '100%'
-}
+const {
+  searchInput: searchValue,
+  showSuggestions,
+  isInlineMode,
+  searchTokens,
+  isExpanded,
+  handleInput,
+  handleFocus,
+  handleBlur,
+  handleClose,
+  handleKeyDown,
+  handleSuggestionClick,
+  removeQualifier
+} = searchInteractions
 
-const unlockScroll = () => {
-  document.body.style.position = ''
-  document.body.style.top = ''
-  document.body.style.width = ''
-  window.scrollTo(0, originalPosition.value)
-}
+// 4. Finalmente las computed properties que dependen de todo lo anterior
+const coloredParts = computed(() => colorText(searchValue.value))
+const suggestions = computed(() => {
+  const usedQualifiers = searchTokens.value
+    .filter(token => token.type === 'qualifier')
+    .map(token => ({
+      qualifier: token.qualifier + ':',
+      value: token.value
+    }))
 
-// Manejadores de eventos
-const handleFocus = () => {
-  isInputFocused.value = true
-  showSuggestions.value = true
-  
-  if (!isInlineMode.value) {
-    isExpanded.value = true
-    lockScroll()
-    nextTick(() => {
-      input.value?.focus()
-    })
-  }
-}
-
-const closeSearch = () => {
-  isInputFocused.value = false
-  showSuggestions.value = false
-  isExpanded.value = false
-  unlockScroll()
-}
-
-const handleBlur = (event: FocusEvent) => {
-  const dialogContent = document.querySelector('.search-dialog-content')
-  if (dialogContent?.contains(event.relatedTarget as Node)) {
-    return
-  }
-  
-  if (!event.relatedTarget || !dialogContent?.contains(event.relatedTarget as Node)) {
-    closeSearch()
-  }
-}
-
-const handleClose = () => {
-  closeSearch()
-}
-
-const handleKeyDown = (event: KeyboardEvent) => {
-  if (event.key === 'Enter') {
-    emit('search')
-    if (!isInlineMode.value) {
-      closeSearch()
-    }
-  } else if (event.key === 'Escape' && isExpanded.value) {
-    closeSearch()
-  }
-  updateCursorPosition()
-}
-
-const handleInput = (event: Event) => {
-  const value = (event.target as HTMLInputElement).value
-  searchInput.value = value
-  emit('update:modelValue', value)
-  nextTick(() => {
-    updateCursorPosition()
-  })
-}
-
-// Modificar handleSuggestionClick para agregar el qualifier correctamente
-const handleSuggestionClick = (qualifier: string) => {
-  const currentValue = searchInput.value
-  // Añadir espacio si el valor actual no termina en espacio
-  const space = currentValue && !currentValue.endsWith(' ') ? ' ' : ''
-  searchInput.value = currentValue + space + qualifier
-  emit('update:modelValue', searchInput.value)
-  
-  nextTick(() => {
-    input.value?.focus()
-    updateCursorPosition()
-    showSuggestions.value = true // Mantener las sugerencias visibles
-  })
-}
-
-// Modificar la función handleSuggestionRemoveClick para coincidir con la firma esperada
-const handleSuggestionRemoveClick = (qualifier: string, value: string) => {
-  const fullQualifier = qualifier.endsWith(':') ? qualifier : `${qualifier}:`
-  const pattern = new RegExp(`${fullQualifier}\\s*${value}\\s*`, 'i')
-  searchInput.value = searchInput.value.replace(pattern, '').trim()
-  emit('update:modelValue', searchInput.value)
-  nextTick(() => {
-    input.value?.focus()
-    updateCursorPosition()
-  })
-}
-
-// Agregar el método handleSearch faltante
-const handleSearch = () => {
-  emit('search')
-  if (!isInlineMode.value) {
-    closeSearch()
-  }
-}
-
-// Keyboard shortcuts
-useSearchShortcuts(handleFocus, () => input.value)
-
-// Sincronización con v-model
-watch(() => props.modelValue, (newValue) => {
-  if (newValue !== searchInput.value) {
-    searchInput.value = newValue
-  }
+  return Object.keys(QUALIFIERS).map(qualifier => ({
+    qualifier,
+    isUsed: usedQualifiers.some(used => used.qualifier === qualifier),
+    value: usedQualifiers.find(used => used.qualifier === qualifier)?.value
+  }))
 })
 
-watch(searchInput, (newValue) => {
-  emit('update:modelValue', newValue)
+// Watchers
+watch(() => props.modelValue, (newValue) => {
+  if (newValue !== searchValue.value) {
+    searchValue.value = newValue
+  }
 })
 </script>
 
 <template>
   <div class="relative w-full">
+    <!-- Normal search input -->
     <div 
-      class="relative flex items-center w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background transition-all duration-200"
-      :class="{ 
-        'ring-2 ring-ring ring-offset-2': showSuggestions,
-        'search-input-expanded': isExpanded 
-      }"
+      ref="searchContainer"
+      class="relative w-full cursor-pointer"
+      :class="{ 'search-active': isExpanded }"
+      @click="handleFocus" 
     >
-      <Icon 
-        name="octicon:search-16"
-        class="mr-2 h-4 w-4 shrink-0 opacity-50"
-      />
-
-      <!-- Input container -->
-      <div class="relative flex-1 overflow-hidden">
-        <div class="relative w-full overflow-x-auto search-scroll-container">
-          <!-- Capa de resaltado con contenido -->
-          <div class="search-highlight absolute inset-0 whitespace-pre pointer-events-none flex items-center">
+      <div 
+        class="h-9 relative flex items-center w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background transition-all duration-200"
+        :class="{ 
+          'ring-2 ring-ring ring-offset-2': showSuggestions,
+          'search-input-expanded': isExpanded 
+        }"
+      >
+        <Icon 
+          name="octicon:search-16"
+          class="mr-2 h-4 w-4 shrink-0 opacity-50"
+        />
+        
+        <!-- Simplificar el contenedor del input -->
+        <div class="relative flex-1">
+          <!-- Colored preview -->
+          <div 
+            aria-hidden="true"
+            class="search-input-mirror absolute inset-0 pointer-events-none flex items-center overflow-hidden"
+          >
             <template v-for="(part, index) in coloredParts" :key="index">
-              <template v-if="part.type === 'qualifier'">
-                <span class="text-primary font-medium">{{ part.text }}</span>
-              </template>
-              <template v-else-if="part.type === 'value'">
-                <span class="search-badge">{{ part.text }}</span>
-              </template>
-              <template v-else>
-                <span>{{ part.text }}</span>
-              </template>
+              <span 
+                v-if="part.type === 'value'" 
+                class="search-badge"
+              >{{ part.text }}</span>
+              <span
+                v-else
+                :class="{
+                  'text-primary font-medium': part.type === 'qualifier',
+                }"
+              >{{ part.text }}</span>
             </template>
-            <!-- Espacio para el cursor -->
-            <span class="pl-px"></span>
+            <!-- Agregar el cursor aquí también -->
+            <span v-if="isInputFocused" class="typing-cursor"></span>
           </div>
 
-          <!-- Input real -->
           <input
             ref="input"
-            v-model="searchInput"
+            :value="searchValue"
             type="text"
-            class="search-input-field relative w-full"
-            :class="{ 'has-content': searchInput }"
+            class="search-input w-full relative"
+            :class="{ 
+              'has-content': searchValue,
+              'text-transparent': searchValue
+            }"
             placeholder="Search repositories..."
             @input="handleInput"
-            @keydown="updateCursorPosition"
-            @keyup="updateCursorPosition"
-            @click="updateCursorPosition"
+            @keydown="handleKeyDown"
             @focus="handleFocus"
             @blur="handleBlur"
+            @scroll="handleInputScroll"
           />
-
-          <!-- Cursor animado -->
-          <span 
-            v-if="isInputFocused" 
-            class="typing-cursor absolute"
-            :style="{ 
-              left: `${cursorPosition}px`,
-              top: '50%',
-              transform: 'translateY(-50%)'
-            }"
-          ></span>
         </div>
+
+        <kbd 
+          v-if="!showSuggestions"
+          class="pointer-events-none hidden h-5 select-none items-center gap-1 rounded border bg-muted px-1.5 font-mono text-[10px] font-medium opacity-100 sm:flex"
+        >
+          <span class="text-xs">/</span>
+        </kbd>
       </div>
 
-      <kbd 
-        v-if="!showSuggestions"
-        class="pointer-events-none hidden h-5 select-none items-center gap-1 rounded border bg-muted px-1.5 font-mono text-[10px] font-medium opacity-100 sm:flex"
+      <!-- Inline suggestions dropdown -->
+      <div
+        v-if="isInlineMode && showSuggestions && suggestions?.length > 0"
+        class="suggestions-dropdown-inline"
       >
-        <span class="text-xs">/</span>
-      </kbd>
-    </div>
-
-    <!-- Suggestions (inline mode) -->
-    <SearchSuggestions
-      v-if="isInlineMode && showSuggestions"
-      :suggestions="suggestions"
-      mode="inline"
-      @select="handleSuggestionClick"
-      @remove="handleSuggestionRemoveClick"
-    />
-
-    <!-- Search Dialog (modal mode) -->
-    <SearchDialog
-      v-if="!isInlineMode && isExpanded"
-      v-model="searchInput"
-      :suggestions="suggestions"
-      @close="handleClose"
-      @search="handleSearch"
-    >
-      <template #search-input>
-        <div class="relative flex-1 overflow-hidden">
-          <div class="relative w-full overflow-x-auto search-scroll-container">
-            <SearchHighlight :tokens="tokens" />
-            <input
-              ref="input"
-              v-model="searchInput"
-              type="text"
-              class="search-input-field relative w-full"
-              :class="{ 'has-content': searchInput }"
-              placeholder="Search repositories..."
-              @input="updateCursorPosition"
-              @keydown="handleKeyDown"
+        <div class="px-2 py-1 text-xs text-muted-foreground font-medium">
+          Available filters
+        </div>
+        <div 
+          v-for="suggestion in suggestions"
+          :key="suggestion.qualifier"
+          class="suggestion-item"
+          :class="{ 'is-used': suggestion.isUsed }"
+          @mousedown.prevent="suggestion.isUsed ? 
+            removeQualifier(suggestion.qualifier, suggestion.value!) : 
+            handleSuggestionClick(suggestion.qualifier)"
+        >
+          <Icon 
+            :name="QUALIFIERS[suggestion.qualifier as keyof typeof QUALIFIERS].icon" 
+            class="h-3.5 w-3.5 opacity-50" 
+          />
+          <span class="suggestion-label text-xs">
+            {{ QUALIFIERS[suggestion.qualifier as keyof typeof QUALIFIERS].label }}
+          </span>
+          <div class="ml-auto flex items-center gap-2">
+            <span v-if="suggestion.isUsed" class="text-xs text-muted-foreground">
+              {{ suggestion.value }}
+            </span>
+            <span class="text-xs text-muted-foreground">
+              {{ suggestion.qualifier }}
+            </span>
+            <Icon 
+              v-if="suggestion.isUsed"
+              name="octicon:x-16"
+              class="h-3.5 w-3.5 opacity-50" 
             />
           </div>
         </div>
-      </template>
-    </SearchDialog>
+      </div>
+    </div>
+
+    <!-- Modal overlay (solo se muestra en modo modal) -->
+    <Transition name="search-overlay">
+      <div 
+        v-if="!isInlineMode && isExpanded" 
+        class="search-overlay"
+        @click.self="handleClose"
+      >
+        <div 
+          class="search-dialog"
+          @click.stop
+        >
+          <div class="search-dialog-content">
+            <div class="flex items-center justify-between mb-4">
+              <h2 class="text-lg font-semibold">Search repositories</h2>
+              <button 
+                class="p-1 rounded-md hover:bg-accent"
+                type="button"
+                @click.stop="handleClose"
+              >
+                <Icon name="octicon:x-16" class="h-4 w-4" />
+              </button>
+            </div>
+            
+            <!-- Search input when expanded -->
+            <div class="relative flex items-center w-full rounded-md border border-input bg-background px-3 py-2 text-sm">
+              <Icon 
+                name="octicon:search-16"
+                class="mr-2 h-4 w-4 shrink-0 opacity-50"
+              />
+              
+              <div class="relative flex-1">
+                <!-- Colored preview and cursor -->
+                <div 
+                  aria-hidden="true"
+                  class="search-input-mirror absolute inset-0 pointer-events-none flex items-center whitespace-pre overflow-x-auto"
+                >
+                  <template v-for="(part, index) in coloredParts" :key="index">
+                    <span v-if="part.type === 'value'" class="search-badge whitespace-pre">{{ part.text }}</span>
+                    <span
+                      v-else
+                      :class="{
+                        'text-primary font-medium whitespace-pre': part.type === 'qualifier',
+                      }"
+                    >{{ part.text }}</span>
+                  </template>
+                  <span v-if="isInputFocused" class="typing-cursor"></span>
+                </div>
+
+                <input
+                  ref="input"
+                  :value="searchValue"
+                  type="text"
+                  class="search-input w-full bg-transparent outline-none relative"
+                  :class="{ 'text-transparent': searchValue }"
+                  placeholder="Search repositories..."
+                  @input="handleInput"
+                  @keydown="handleKeyDown"
+                  @scroll="handleInputScroll"
+                />
+              </div>
+            </div>
+
+            <!-- Suggestions in dialog -->
+            <div v-if="suggestions?.length > 0" class="mt-2">
+              <div class="px-2 mb-2 text-xs text-muted-foreground font-medium">
+                Available filters
+              </div>
+              <div 
+                v-for="suggestion in suggestions"
+                :key="suggestion.qualifier"
+                class="suggestion-item"
+                :class="{ 'is-used': suggestion.isUsed }"
+                @mousedown.prevent="suggestion.isUsed ? 
+                  removeQualifier(suggestion.qualifier, suggestion.value!) : 
+                  handleSuggestionClick(suggestion.qualifier)"
+              >
+                <Icon 
+                  :name="QUALIFIERS[suggestion.qualifier as keyof typeof QUALIFIERS].icon" 
+                  class="h-3.5 w-3.5 opacity-50" 
+                />
+                <span class="suggestion-label text-xs">
+                  {{ QUALIFIERS[suggestion.qualifier as keyof typeof QUALIFIERS].label }}
+                </span>
+                <div class="ml-auto flex items-center gap-2">
+                  <span v-if="suggestion.isUsed" class="text-xs text-muted-foreground">
+                    {{ suggestion.value }}
+                  </span>
+                  <span class="text-xs text-muted-foreground">
+                    {{ suggestion.qualifier }}
+                  </span>
+                  <Icon 
+                    v-if="suggestion.isUsed"
+                    name="octicon:x-16"
+                    class="h-3.5 w-3.5 opacity-50" 
+                  />
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </Transition>
   </div>
 </template>
 
 <style scoped>
-.search-scroll-container {
-  position: relative; /* Asegura que el contenedor sea relativo */
+.suggestions-dropdown {
+  position: absolute;
+  top: calc(100% + 0.5rem);
+  left: 0;
+  right: 0;
+  background-color: hsl(var(--background));
+  border: 1px solid hsl(var(--border));
+  border-radius: 0.375rem;
+  box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1), 0 2px 4px -2px rgb(0 0 0 / 0.1);
+  z-index: 50;
+  padding: 0.5rem 0;
+}
+
+.suggestion-item {
+  padding: 0.5rem;
   display: flex;
   align-items: center;
-  overflow-x: auto;
-  scrollbar-width: none;
-  -ms-overflow-style: none;
-  margin-right: -8px;
-  padding-right: 8px;
+  gap: 0.625rem;
+  cursor: pointer;
+  border-radius: 0.375rem;
+  transition: background-color 0.2s ease;
+  font-size: 0.75rem;
+  line-height: 1rem;
 }
 
-.search-scroll-container::-webkit-scrollbar {
-  display: none;
+.suggestion-item:hover {
+  background-color: hsl(var(--accent));
 }
 
-.search-input-field {
-  position: relative;
-  background: transparent;
-  outline: none;
-  color: transparent;
-  caret-color: transparent; /* Ocultar el cursor nativo */
-  min-width: 100%;
-  white-space: pre;
-  padding: 2px 0;
+.suggestion-item.is-used {
+  background-color: hsl(var(--muted)/0.1);
 }
 
-.search-input-field.has-content {
-  -webkit-text-fill-color: transparent;
+.suggestion-item.is-used:hover {
+  background-color: hsl(var(--muted)/0.2);
 }
 
-.search-input-field::placeholder {
-  -webkit-text-fill-color: initial;
-  color: hsl(var(--muted-foreground));
+.suggestion-label {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+/* ...rest of existing styles... */
+.text-transparent {
+  -webkit-text-fill-color: currentColor;
 }
 
 .typing-cursor {
-  position: absolute;
+  display: inline-block;
   width: 2px;
   height: 1.2em;
   background-color: currentColor;
-  pointer-events: none;
+  margin-left: 1px;
   animation: blink 1s step-end infinite;
-  z-index: 10; /* Asegura que el cursor esté siempre visible */
-  margin-left: -1px; /* Ajuste fino de la posición */
+  position: relative;
+  top: 1px;
 }
 
 @keyframes blink {
-  0%, 100% { opacity: 1; }
-  50% { opacity: 0; }
+  0%, 100% {
+    opacity: 1;
+  }
+  50% {
+    opacity: 0;
+  }
 }
 
 .search-badge {
@@ -338,12 +384,72 @@ watch(searchInput, (newValue) => {
   line-height: 1.25rem;
   font-weight: 500;
   margin: 0 1px;
-  white-space: pre;
 }
 
 :root[class~="dark"] .search-badge {
   background-color: #1a384d;
   color: #2f81f7;
+}
+
+.search-expanded {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  height: 3.5rem; /* Match header height */
+  z-index: 50;
+  display: flex;
+  align-items: center;
+  background-color: hsl(var(--background));
+  border-bottom: 1px solid hsl(var(--border));
+  backdrop-filter: blur(8px);
+}
+
+.search-input-expanded {
+  max-width: 60rem;
+  margin: 0 auto;
+  width: 100%;
+  position: relative;
+}
+
+.suggestions-dropdown {
+  position: fixed;
+  top: 3.5rem; /* Position right below the header */
+  left: 50%;
+  transform: translateX(-50%);
+  width: 60rem !important;
+  max-width: calc(100vw - 2rem);
+  margin-top: 0;
+  background-color: hsl(var(--background));
+  border: 1px solid hsl(var(--border));
+  border-radius: 0.375rem;
+  box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1), 0 2px 4px -2px rgb(0 0 0 / 0.1);
+}
+
+/* Ensure proper stacking */
+:root {
+  --search-z-index: 100; /* Increased to appear above header */
+}
+
+.search-expanded {
+  z-index: var(--search-z-index);
+}
+
+.suggestions-dropdown {
+  z-index: calc(var(--search-z-index) + 1);
+}
+
+/* Add backdrop blur for better visibility */
+.search-expanded {
+  background-color: hsl(var(--background)/95%);
+  backdrop-filter: blur(8px);
+}
+
+/* Add smooth transitions */
+.search-expanded,
+.search-input-expanded,
+.suggestions-dropdown {
+  transition: all 0.2s ease-out;
 }
 
 .search-overlay {
@@ -370,5 +476,404 @@ watch(searchInput, (newValue) => {
   border-radius: 0.75rem;
   padding: 1.5rem;
   box-shadow: 0 20px 25px -5px rgb(0 0 0 / 0.1), 0 8px 10px -6px rgb(0 0 0 / 0.1);
+}
+
+/* Transition animations */
+.search-overlay-enter-active,
+.search-overlay-leave-active {
+  transition: opacity 0.2s ease;
+}
+
+.search-overlay-enter-from,
+.search-overlay-leave-to {
+  opacity: 0;
+}
+
+.search-overlay-enter-active .search-dialog,
+.search-overlay-leave-active .search-dialog {
+  transition: transform 0.15s ease-out;
+}
+
+.search-overlay-enter-from .search-dialog,
+.search-overlay-leave-to .search-dialog {
+  transform: translateY(-1rem);
+}
+
+/* Asegurar que el diálogo permanezca visible */
+.search-dialog {
+  transform: translateY(0);
+}
+
+.search-active {
+  @apply invisible;
+}
+
+.suggestions-dropdown-inline {
+  position: absolute;
+  top: calc(100% + 0.25rem);
+  left: 0;
+  right: 0;
+  background-color: hsl(var(--background));
+  border: 1px solid hsl(var(--border));
+  border-radius: 0.375rem;
+  box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1), 0 2px 4px -2px rgb(0 0 0 / 0.1);
+  z-index: 50;
+  padding: 0.5rem 0;
+  max-height: 16rem;
+  overflow-y: auto;
+}
+
+/* Asegurar que el input mantenga el texto visible */
+.text-transparent {
+  -webkit-text-fill-color: transparent;
+  color: transparent;
+  caret-color: currentColor; /* Mantener el cursor visible */
+}
+
+.search-input-mirror {
+  overflow-x: auto;
+  white-space: pre;
+  scrollbar-width: none; /* Firefox */
+  -ms-overflow-style: none; /* IE and Edge */
+}
+
+.search-input-mirror::-webkit-scrollbar {
+  display: none; /* Chrome, Safari and Opera */
+}
+
+.search-input {
+  white-space: pre;
+  overflow-x: auto;
+}
+
+/* Ajustar el estilo del texto transparente para mejor visibilidad */
+.text-transparent {
+  -webkit-text-fill-color: transparent;
+  color: transparent;
+}
+
+/* Asegurar que los badges mantengan su espacio */
+.search-badge {
+  flex-shrink: 0;
+  display: inline-flex;
+  white-space: pre;
+}
+
+/* Asegurar que el contenedor mantenga el scroll sincronizado */
+.relative.flex-1 {
+  overflow: hidden;
+}
+
+.search-wrapper {
+  position: relative;
+  overflow-x: auto;
+  scrollbar-width: none;
+  -ms-overflow-style: none;
+}
+
+.search-wrapper::-webkit-scrollbar {
+  display: none;
+}
+
+.search-mirror {
+  position: absolute;
+  top: 0;
+  left: 0;
+  padding: inherit;
+  pointer-events: none;
+  visibility: visible;
+  white-space: pre;
+  display: flex;
+  align-items: center;
+}
+
+.search-input {
+  width: 100%;
+  background: transparent;
+  outline: none;
+  color: transparent;
+  caret-color: currentColor;
+  position: relative;
+  white-space: pre;
+  padding: inherit;
+}
+
+.search-input.has-content {
+  -webkit-text-fill-color: transparent;
+}
+
+.search-input::placeholder {
+  -webkit-text-fill-color: initial;
+  color: initial;
+}
+
+.search-badge {
+  display: inline-flex;
+  align-items: center;
+  background-color: #ddf4ff;
+  color: #0969da;
+  border-radius: .2rem;
+  padding: 0 0.5rem;
+  font-size: 0.875rem;
+  line-height: 1.25rem;
+  font-weight: 500;
+  margin: 0 1px;
+  white-space: pre;
+}
+
+:root[class~="dark"] .search-badge {
+  background-color: #1a384d;
+  color: #2f81f7;
+}
+
+.search-container {
+  position: relative;
+  min-width: 0;
+  width: 100%;
+}
+
+.search-content-wrapper {
+  position: relative;
+  width: 100%;
+  overflow-x: auto;
+  -ms-overflow-style: none;
+  scrollbar-width: none;
+}
+
+.search-content-wrapper::-webkit-scrollbar {
+  display: none;
+}
+
+.search-mirror {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  pointer-events: none;
+  white-space: pre;
+  padding: 2px 0;
+  min-width: min-content;
+  display: flex;
+  align-items: center;
+}
+
+.search-input {
+  width: 90%;
+  background: transparent;
+  outline: none;
+  position: relative;
+  white-space: pre;
+  padding: 2px 0;
+  min-width: 90%;
+  color: transparent;
+  caret-color: currentColor;
+}
+
+.search-input::placeholder {
+  color: hsl(var(--muted-foreground));
+  -webkit-text-fill-color: initial;
+}
+
+.search-badge {
+  display: inline-flex;
+  align-items: center;
+  background-color: #ddf4ff;
+  color: #0969da;
+  border-radius: .2rem;
+  padding: 0 0.5rem;
+  font-size: 0.875rem;
+  line-height: 1.25rem;
+  font-weight: 500;
+}
+
+.search-input-container {
+  position: relative;
+  width: 100%;
+  overflow: hidden;
+}
+
+.search-input-wrapper {
+  position: relative;
+  width: 100%;
+}
+
+.search-highlight-layer {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  pointer-events: none;
+  white-space: pre;
+  overflow: hidden;
+  display: flex;
+  align-items: center;
+  gap: 0.25rem; /* Nuevo: separación consistente entre elementos */
+}
+
+.search-input-field {
+  width: 100%;
+  background: transparent;
+  outline: none;
+  color: transparent;
+  caret-color: currentColor;
+  white-space: pre;
+}
+
+.search-input-field::placeholder {
+  color: hsl(var(--muted-foreground));
+  -webkit-text-fill-color: initial;
+}
+
+.search-badge {
+  display: inline-flex;
+  align-items: center;
+  padding: 0 0.25rem;
+  border-radius: 0.2rem;
+  background: #ddf4ff;
+  color: #0969da;
+  font-size: 0.875rem;
+  font-weight: 500;
+  white-space: pre;
+  /* Se puede eliminar o reducir "margin" ya que el gap se encarga */
+}
+
+:root[class~="dark"] .search-badge {
+  background: #1a384d;
+  color: #2f81f7;
+}
+
+.extra-space {
+  /* Forzar al menos 2 caracteres de espacio extra */
+  min-width: 2ch;
+  flex-shrink: 0;
+}
+
+/* Remove duplicate styles and keep only these essential ones */
+.search-scroll-container {
+  display: flex;
+  align-items: center;
+  overflow-x: auto;
+  scrollbar-width: none;
+  -ms-overflow-style: none;
+  margin-right: -8px; /* Compensar el padding del contenedor */
+  padding-right: 8px;
+}
+
+.search-scroll-container::-webkit-scrollbar {
+  display: none;
+}
+
+/* Ajustar el input */
+.search-input-field {
+  background: transparent;
+  outline: none;
+  color: transparent;
+  caret-color: currentColor;
+  min-width: 100%;
+  position: relative;
+  white-space: pre;
+  padding: 2px 0;
+}
+
+.search-input-field.has-content {
+  -webkit-text-fill-color: transparent;
+}
+
+.search-input-field::placeholder {
+  -webkit-text-fill-color: initial;
+  color: hsl(var(--muted-foreground));
+}
+
+/* Ajustar la capa de resaltado */
+.search-highlight {
+  display: flex;
+  align-items: center;
+  white-space: pre;
+  padding: 2px 0;
+}
+
+/* Asegurar que los badges mantengan su forma */
+.search-badge {
+  display: inline-flex;
+  align-items: center;
+  white-space: pre;
+  flex-shrink: 0;
+}
+
+/* Ajustar el cursor */
+.typing-cursor {
+  width: 2px;
+  height: 1.2em;
+  background-color: currentColor;
+  pointer-events: none;
+}
+
+.suggestions-group + .suggestions-group {
+  margin-top: 0.5rem;
+}
+
+.search-badge-mini {
+  background-color: hsl(var(--accent));
+  color: hsl(var(--accent-foreground));
+  padding: 0 0.375rem;
+  border-radius: 0.25rem;
+  font-size: 0.75rem;
+  line-height: 1.25rem;
+}
+
+.suggestion-item.is-used {
+  background-color: hsl(var(--muted)/0.1);
+}
+
+.suggestion-item.is-used:hover {
+  background-color: hsl(var(--destructive)/0.1);
+}
+
+.suggestions-container {
+  background-color: hsl(var(--background));
+  border: 1px solid hsl(var(--border));
+  border-radius: 0.375rem;
+  box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1), 0 2px 4px -2px rgb(0 0 0 / 0.1);
+  max-height: 24rem;
+  overflow-y: auto;
+}
+
+.suggestion-item.is-used:hover {
+  background-color: hsl(var(--destructive)/0.1);
+  color: hsl(var(--destructive));
+}
+
+.suggestion-item.is-used:hover .search-badge-mini {
+  background-color: hsl(var(--destructive)/0.2);
+  color: hsl(var(--destructive));
+}
+
+.search-input-field {
+  background: transparent;
+  outline: none;
+  position: relative;
+  white-space: pre;
+  padding: 2px 0;
+  caret-color: currentColor; /* Asegura que el cursor sea visible */
+  color: currentColor; /* Color del texto igual al tema actual */
+}
+
+/* Solo hacer el texto transparente cuando tiene contenido y no está en modo inline */
+.search-input-field.has-content:not(.inline-mode) {
+  color: transparent;
+  -webkit-text-fill-color: transparent;
+}
+
+/* Asegurar que el cursor sea visible en modo inline */
+.search-input-field.inline-mode {
+  color: currentColor;
+  -webkit-text-fill-color: currentColor;
+}
+
+/* Mantener el placeholder visible */
+.search-input-field::placeholder {
+  -webkit-text-fill-color: initial;
+  color: hsl(var(--muted-foreground));
 }
 </style>
